Subclassed NeuralNet as a base class of NeuralNetBase, and added new lockfree network implementation with new sort development in progress.
