+++ libNeat/include/NeuralNet.hh
@@ -2,78 +2,35 @@
 #include <vector>
 #include <stdexcept>
 #include <functional>
+#include "NeuralNetBase.hh"
 
 // A network is the current real world view of an organisms
 // internal network. It is what is used for performing calc-
 // ulations and represents the state of all enabled links of
 // a genome neuron set. Referred to as a pheonotype by NEAT.
 
-struct NodeGene;
-
-enum class NodeType { Input, Hidden, Output, Bias };
-bool IsSensor(const NodeType& type);
-
-struct Node {
-  double value;
-  bool is_sigmoid;
-  NodeType type;
-  Node(NodeType _type)
-    : value(0.0),
-      is_sigmoid(false), type(_type) {;}
-};
-
-enum class ConnectionType { Normal, Recurrent };
-
-struct Connection {
-  unsigned int origin;
-  unsigned int dest;
-  double weight;
-  ConnectionType type;
-  Connection(unsigned int _origin,
-             unsigned int _dest,
-             ConnectionType _type,
-             double _weight=0.5)
-    : origin(_origin), dest(_dest),
-      weight(_weight), type(_type) {;}
-};
-
-class NeuralNet {
+class NeuralNet : public NeuralNetBase {
+  friend class LFNetwork;
 public:
-  NeuralNet(std::vector<Node>&& Nodes, std::vector<Connection>&& Conn);
-  NeuralNet(const std::vector<NodeGene>& Nodes);
+  using NeuralNetBase::NeuralNetBase;
 
   std::vector<double> evaluate(std::vector<double> inputs);
-  void load_input_vals(const std::vector<double>& inputs);
-  std::vector<double> read_output_vals();
-  void add_connection(int origin, int dest, double weight);
-  bool would_make_loop(unsigned int i, unsigned int j) const;
-  void register_sigmoid(std::function<double(double)> sig) {sigma = sig;}
-
-  unsigned int num_nodes() const { return nodes.size(); }
-  unsigned int num_connections() const { return connections.size(); }
-
-  std::vector<NodeType> node_types() const;
-  const std::vector<Connection>& get_connections() const { return connections; }
 
 private:
-  double sigmoid(double val) const;
-  double get_node_val(unsigned int i);
-  void add_to_val(unsigned int i, double val);
-  void sort_connections();
+  void sort_connections() override;
 
-  std::vector<Node> nodes;
-  std::vector<Connection> connections;
-  bool connections_sorted;
-  std::function<double(double val)> sigma;
-
-  friend std::ostream& operator<<(std::ostream& os, const NeuralNet& net);
 };
 
 
-class NetworkException : public std::exception {
-  using std::exception::exception;
-};
+class LFNetwork : public NeuralNetBase {
+public:
+  using NeuralNetBase::NeuralNetBase;
+  LFNetwork(NeuralNet&& net);
 
-class NetworkSensorSize : public NetworkException {
-  using NetworkException::NetworkException;
+  void sort_connections() override;
+
+
+private:
+  enum class EvaluationOrder { GreaterThan, LessThan, NotEqual, Unknown };
+  EvaluationOrder compare_connections(const Connection& a, const Connection& b);
 };
